refactor: Setting up the app state listener now happens once on app start, by using context and moving logic
